add versioning

+++ src/connector/posix_connector.cpp
@@ -231,7 +231,7 @@ int PosixE3Connector::receive(std::vector<uint8_t>& buffer) {
         return static_cast<int>(ErrorCode::TRANSPORT_ERROR);
     }
     
-    buffer.resize(ret);
+    buffer.resize(static_cast<size_t>(ret));
     E3_LOG_TRACE(LOG_TAG) << "Received: " << ret << " bytes";
     return static_cast<int>(ret);
 }
@@ -367,9 +367,9 @@ int PosixE3Connector::send_in_chunks(int sockfd, const uint8_t* buffer, size_t b
     
     size_t total_sent = 0;
     while (total_sent < buffer_size) {
-        size_t bytes_to_send = std::min(static_cast<size_t>(CHUNK_SIZE), buffer_size - total_sent);
+        ssize_t bytes_to_send = std::min(static_cast<size_t>(CHUNK_SIZE), buffer_size - total_sent);
         
-        size_t chunk_sent = 0;
+        ssize_t chunk_sent = 0;
         while (chunk_sent < bytes_to_send) {
             ssize_t sent_chunk = ::send(sockfd, buffer + total_sent + chunk_sent, 
                                        bytes_to_send - chunk_sent, 0);
